# docs/Jothidam_AI_OpenAPI_v1_Thirukanitham_2026.yaml
openapi: 3.0.3
info:
  title: Vinaadi AI API
  version: 1.0.0
  description: MVP API for Thirukanitham Tamil astrology companion with Family Aggregate Fortune in MVP 1.
servers:
  - url: https://api.jothidam.ai/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: object }
        meta: { type: object }
paths:
  /birth-profiles:
    post:
      summary: Create birth profile for self or family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, birthDateLocal, birthPlace, birthLatitude, birthLongitude, birthTimezone]
              properties:
                ownerUserId: { type: string, format: uuid }
                familyVaultId: { type: string, format: uuid, nullable: true }
                relationshipToOwner: { type: string, enum: [self, spouse, child, parent, sibling, grandparent, other] }
                displayName: { type: string }
                birthDateLocal: { type: string, format: date }
                birthTimeLocal: { type: string, example: '06:30:00', nullable: true }
                birthPlace: { type: string }
                birthLatitude: { type: number }
                birthLongitude: { type: number }
                birthTimezone: { type: string, example: Asia/Kolkata }
                birthTimeConfidenceMinutes: { type: integer }
                calculateNow: { type: boolean, default: true }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } } }
  /birth-profiles/{birthProfileId}:
    get:
      summary: Get birth profile
      parameters:
        - in: path
          name: birthProfileId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Birth profile }
...
  /charts/{chartId}/summary:
    get:
      summary: Get chart summary
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: language
          schema: { type: string, enum: [ta, en, ta-en] }
      responses:
        '200': { description: Chart summary }
...
  /charts/{chartId}/dasha:
    get:
      summary: Get Dasha timeline and active period
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: level
          schema: { type: string, enum: [maha, antar, pratyantar, sookshma, prana] }
        - in: query
          name: asOf
          schema: { type: string, format: date }
      responses:
        '200': { description: Dasha timeline }
...
  /charts/{chartId}/daily-guidance:
    get:
      summary: Get individual daily 0-100 guidance score
      parameters:
        - in: path
          name: chartId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: date
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Daily guidance }
  /daily-guidance/range:
    get:
      summary: Get daily guidance range by birth profile
      parameters:
        - in: query
          name: profileId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: from
          required: true
          schema: { type: string, format: date }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Daily guidance range }
...